+++ Sources/Shared/ProjectKind.swift
@@ -3,4 +3,5 @@ import Foundation
 public enum ProjectKind: String {
     case xcode = "Xcode"
     case spm = "Swift Package Manager"
+    case generic = "Generic"
 }