+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ##### Enhancements
 
 - Add CodeClimate output formatter available via the `--format codeclimate` option. 
+- Add support for third-party build systems, such as Bazel.
 
 ##### Bug Fixes